[FEAT] 기능 추가 및 연관된 로직 병합 (#5)
- 비회원 상수를 프로퍼티에서 Enum 클래스로 변경
- 비회원 검증 로직 방식 변경
- 불필요 로직 일부 삭제

+++ src/main/java/com/project/trysketch/suggest/NounRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 // 1. 기능   : 명사 Repository
 // 2. 작성자 : 서혁수
-public interface NounRepository extends JpaRepository<NounEntity, Long> {
+public interface NounRepository extends JpaRepository<NounEntity, Integer> {
 }
